refactor: jshelper -> jsutil
