+++ plugin/internal/jsutil/transpilation_test.go
@@ -1,4 +1,4 @@
-package jshelper
+package jsutil
 
 import (
 	"testing"
